requirements.txt
absl-py==2.3.0
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
boto3==1.38.24
botocore==1.38.24
certifi==2025.4.26
click==8.2.1
distro==1.9.0
exceptiongroup==1.3.0
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.2
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
immutabledict==4.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
MarkupSafe==3.0.2
numpy==2.2.6
openai==1.82.0
ortools==9.12.4544
packaging==25.0
pandas==2.2.3
protobuf==5.29.4
psycopg2-binary==2.9.10
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
pytz==2025.2
s3transfer==0.13.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
